+++ .github/CHANGELOG.md
@@ -162,6 +162,7 @@
 1. [A380X/FWS] Add RAT OUT, REFUEL PNL DOOR, HI ALT AIRPORT & REFUELING/DEFUELING memos - @BravoMike99 (bruno_pt99)
 1. [A380X/ND] Correct positioning of TCAS Messages and added TCAS STBY - @MrJigs7 (MrJigs)
 1. [A32NX/AFS] Add new vertical profile validity LVar for managed vertical modes - @lukecologne (luke)
+1. [A380X/MFD] Add ATCCOM MSG RECORD page layouts - @heclak (heclak)
 
 ## 0.12.0